move scripts dir, update README
A few changes to make it easier to run scripts on the Xposition server
