+++ xposition/scripts/delete_categories.py
