+++ xposition/scripts/tex2markdown.py
